*: move pkg/colorutil to pkg/tui
